Add tests for different plugin postfixes
