+++ Tests/TechTalk.SpecFlow.GeneratorTests/GeneratorTests.csproj
@@ -72,6 +72,7 @@
   <ItemGroup>
     <Compile Include="ConfigurationTests.cs" />
     <Compile Include="GeneratorPluginLocatorTests.cs" />
+    <Compile Include="MockFileSystem.cs" />
     <Compile Include="TestGeneratorTest.cs" />
     <Compile Include="CodeDomExtensions.cs" />
     <Compile Include="CustomTestGeneratorProviderTest.cs" />
